♻️ extract requests.specHelper
