+++ packages/core/test/index.ts
@@ -2,6 +2,7 @@ export * from './browserChecks'
 export * from './buildEnv'
 export * from './capturedExceptions'
 export * from './collectAsyncCalls'
+export * from './requests.specHelper'
 export * from './stubReportApis'
 export * from './stubZoneJs'
 export * from './syntheticsWorkerValues'
